add super abilities tj work on it
